Adds some missing (necessary) files and revisions.
It looks like something went a little weird when updating to the latest
submodule reference, where some bindings got seemingly jumbled.

The maintainers of Cocos2d-x recommend to use the bindings-generator project in order to
auto-gen these bindings. The project can be found at
https://github.com/cocos2d/bindings-generator.

It's worth noting that I *did not* use the bindings generator because it
requires us to remove the Homebrew package manager and instead use
Macports. Re-creating my environment and hard-coding the bindings became
much easier to deal with.